[6x]: Fix gpcheckperf discrepancy in network test results (v6.23 vs v6.25) (#16389)
Fixes - https://github.com/greenplum-db/gpdb/issues/16356

* Fix gpcheckperf discrepancy in network test results (v6.23 vs v6.25)

Problem - gpcheckperf provides different results when executing a sequential
network test with a specified hostfile between versions 6.23 and 6.25.

RCA - The problem arises from the getHostList function, which assigns GV.opt['-h']
with a list of hosts because of which the code interprets that the -h option is set
with the host list.As GV.opt['-h'] is already assigned during network testing, the
code incorrectly combines hosts from both the GV.opt['-f'] option and GV.opt['-h'],
when it should exclusively retrieve hosts from either GV.opt['-f'] or GV.opt['-h'].
It leads to redundant host list data, resulting in the observed issue

Solution - Omit the assignment of the global variable GV.opt['-h'] within the "getHostList"
function. Instead, utilize a function call to retrieve the host list when it is required in
other parts of the code.

Test - Added behave test case for the fix.