Remove pg_xlogdump -f flag from pg_basebackup test
For one of the pg_basebackup tap test scenarios, we would run pg_xlogdump
with the follow(-f) flag. In some scenarios as mentioned by the previous
commit, it's possible that the latest WAL file may not contain the
SWITCH record and hence might be ready to accept new WAL records. With
the follow flag, we would wait forever for these new WAL records until
the SWITCH wal record showed up which wouldn't happen so the test would
get stuck indefinitely.

We don't really need this flag since the previous pg_basebackup call
would guarantee a SWITCH record in the WAL file. So it's best to remove
this flag.