[6x]: Cleanup orphaned directory of dropped database after differential recovery (#16525)
* Cleanup orphaned directory of the dropped database after differential recovery

Issue - Even if the database is dropped before the recovery process, differential recovery
still leaves behind an orphaned directory for the dropped database under the base directory.

RCA - When a database is created, a file named pg_internal.init is also generated in the
database directory. During the rsync process of pg_data, we intentionally exclude this
pg_internal.init file (--exclude=pg_internal.init). However, because this file remains
in the database directory and is excluded from rsync, it results in an error message
stating "cannot delete non-empty directory: base/db_oid". As a consequence, it becomes
problematic to delete the directory of a dropped database.

Fix - Removed the ‘pg_internal.init’ file from the excluded list of rsync, so that it
won't be problematic to delete the directory of a dropped database. pg_internal.init file
is rebuilt on startup so there is no problem if it is copied during sync_pg_data.