Allow absolute path for pg_file_rename
If caller indicates so. Only for superuser. This is to support utility function
gp_move_orphaned_files which can move files between absolute paths. So far only
pg_file_rename requires it.

P.S. in 7X superuser can do the same but it is achieved via a new role
'pg_read_server_files' which is not in 6X. So adding the superuser() check
instead.