[6X] Fix SET command that sends DTX protocol command when shouldn't
Backport from #16675. No merge conflict.

Original commit message:

If there is an explicit BEGIN, and if the transaction only has SET command, we
will begin transaction and setup DTX context on QEs at the time of the first
SET command. So we need to add dtxSegments in SET to make sure we end the
transaction at the time of END/COMMIT. This is required since b43629b
where we no longer start DTX at the time of BEGIN.

However, we should not do this when there's NO explicit BEGIN. Because in that
case, QEs will not have DTX context setup with the SET command, so if we decide
to send DTX protocol command later (e.g. we have some direct-dispatch command
that requires DTX with part of the QEs), we will send DTX protocol command to
ALL dtxSegments. As a result, some QE might fail to recognize the DTX protocol
command and throw error. E.g.:

	create table t1(a int, b int);

	CREATE OR REPLACE FUNCTION dtx_bug()
	  RETURNS void
	  LANGUAGE plpgsql
	AS $function$
	BEGIN
	    EXECUTE 'UPDATE t1 set b = 1 where a = 1;';
	    set optimizer=off;
	END;
	$function$;

	postgres=# SELECT dtx_bug();
	ERROR:  Distributed transaction 16607 not found (cdbtm.c:2185)  (seg2 127.0.1.1:7004 pid=32375) (cdbtm.c:2185)

Fix the case by skipping adding dtxSegments when w/o explicit BEGIN/END. Also
fix the comment in code and add test cases.