6X: Check peer listener failed in IC-PROXY mode (#16553)
This commit backported #16438 to 6x:

In IC-PROXY mode, the user doesn't get a notification when peer listener bind/listen failed. Because the related code is in IC-PROXY process, it only records warning logs. So, the user's query just hangs here silently. This behavior may make the user very confused.

In the commit, introduced a SHM variable to check the failure and give notification in time.