Retry gang creation for non-recovery failures
This is a backport of 7X commit:
99a48e65b803ae7e446894177f4f97c10415ac18 with one very notable change:
the GUC gp_gang_creation_retry_non_recovery controls whether we want to
retry on non-recovery failures, giving users a chance to opt out of the
behavior.

Original commit message follows:

Before this commit, any error from a segment that is unrelated to
recovery, would result in an immediate termination with the following:

ereport(ERROR, (errcode(ERRCODE_GP_INTERCONNECTION_ERROR),
	errmsg("failed to acquire resources on one or more segments"),
	errdetail("%s (%s)", PQerrorMessage(segdbDesc->conn), segdbDesc->whoami)));

In other words, this termination did not respect
gp_gang_creation_retry_count. From now on, we exhaust the retries first,
before emitting this ERROR.

Co-authored-by: Ashwin Agrawal <aashwin@vmware.com>
Reviewed-by: Xin Zhang <xinzweb@users.noreply.github.com>
Reviewed-by: Huansong Fu <fuhuansong@gmail.com>