Fix gpcheckcat 'oid does not exist' error for catalog tables
**Issue:**
gpcheckcat reports `column "oid" does not exist` error while running ‘missing_extraneous’ test on pg_shdescription catalogue table.
This is because internally 'missing_extraneous' test of gpcheckcat tries to fetch the oid of all the catalogue tables present in the database.
For pg_shdescription table, the oid was changed to objoid in 6x and 7x.
So there is no attribute as oid in pg_shdescription table.
Similar issue will be seen for other catalogue tables wherever the 'attname' field of pg_attribute table is objoid.
These tables are as below:
- pg_seclabels
- pg_description
- pg_shdescription
- pg_seclabel
- pg_shseclabel

Issue with pg_description table is fixed as part of [#16140 ](https://github.com/greenplum-db/gpdb/pull/16140)
This PR is used to track the fix for other catalog tables listed above.

**Fix:**
Catalog tables which do not have a OID column are stored in a dictionary 'TableMainColumn' and are mapped an existing OID type attribute. An entry for the required catalog tables is added to dictionary so that objoid is used instead of oid while running SQL queries.

**Test:**
A behave test is added for pg_shdescription table.