Upgrade pgbouncer to 1.21.0 tag for GPDB6. (#16863)
The upstream of pgbouncer has fixed a lot of crash issues. So it is better to upgrade pgbouncer to the latest version before customers reporting some crash issues in the future.
Also the latest branch of pgbouncer has set submodule libusual to upstream repo:https://github.com//libusual/libusual.
And another submodule is added to pgbouncer which is uthash to repo:https://github.com/troydhanson/uthash.git.