Fix memory leak in Dynamic Index/BitmapIndex Scan during execution (#16797) (#16851)
This fix addresses a memory leak that occurs during the execution of DynamicIndexScan
and DynamicBitmapIndexScan. Each partition of the DynamicIndexScan creates an IndexScanState
for execution usage. Within each IndexScanState, an ExprContext is allocated but not freed
in ExecFreeExprContext(). This ExprContext includes a MemoryContext, resulting in a memory
leak of the initial context space during each iteration of the DynamicIndexScan partition loop.
Which will lead to a significant accumulation of leaked memory if the partition loop is extensive.
The same issue is observed in the execution of DynamicBitmapIndexScan.

This fix ensures that the allocated MemoryContext within the ExprContext is properly freed during
the execution of each partition in Dynamic Index/BitmapIndex Scan.

Authored-by: wuyuhao28 <wuyuhao28@github.com>