[6X] pg_dump: fix --function-oid when --relation-oid is also used
This is a backport of 7X commits 992b4dd and 9a36a39 with minor
modifications to account for differences in pg_dump code between
6X and 7X.

Original commit message follows:

pg_dump does not correctly dump functions when both greenplum specific
internal flags `--function-oid` and `--relation-oid` are used at the
same time. This is because functions are being marked dumpable or not
twice. The first time is by selectDumpableFunction and then again by
selectDumpableObject. When `--relation-oid` is used, all namespaces are
marked not dumpable which is why selectDumpableObject will mark the
function to not be dumped.

This must be a regression. The original commit that introduces greenplum
specific selectDumpableFunction removes the usage of the more generic
selectDumpableObject.

selectDumpableFunction reference commit:
318b86a

selectDumpableAggregate was created so aggregate dumping works as intended when
using --function-oid on an aggregate function. This is necessary because
pg_dump treats aggregates separately from functions. See the comment in getFuncs
for details.