Fix the bug which leading to wrong answer of sql string_agg(orderby) & agg(distinct).
GP makes optimization for distinct. When each agg is initially grouped, 
some sortstate need to be deduplicated which is belong to distinct , 
which can reduce the memory footprint of the sorted sortstate and
improve the performance of deduplicating in the later stage. However,
due to the mistake of the code, it may occur that when there are multiple
agg in the SQL, and one of them needs to be deduplicated, but more than
one agg needs to be sorted, all sortstate are identified as need to be
deduplicated. This will affect the results of non-distinct agg, because it is
mistakenly overweight, so from this point of view, this is actually a
general problem.

Fix Github Issue #17028

Authored-by: erchuan.ty <erchuan.ty@alibaba-inc.com>
Co-authored-by: Zhenghua Lyu <kainwen@gmail.com>