Updated help doc in consistent with code (#16896) (#16911)
Issue: gpexpand --help was showing wrong for -n option

-n <parallel_processes>
The number of tables to redistribute simultaneously. Valid values
are 1 - 16. Each table redistribution process requires two database
connections: one to alter the table, and another to update the table's
status in the expansion schema. Before increasing -n, check the current
value of the server configuration parameter max_connections and make
sure the maximum connection limit is not exceeded.
instead of 1 - 16 it should be 1 - 96 as per the code MAX_PARALLEL_EXPANDS = 96

Fix: updated max range value in the help file.