Handle ICProxy addresses during gpexpand (#17009)
This commit ported #16861 to 6x.

----
Current gpexpand doesn't handle the ICProxy addresses. So, if we are in icproxy mode: the proxy addresses of newly added segments are still empty after expansion. This causes queries to hang forever.

We have to check if the user has already set gp_interconnect_proxy_addresses before and prompt them to set a proper value via gpconfig before expansion.