pg_upgrade: add support func to check if view contains removed types
Views that use removed types will cause upgrade to fail. This happens
during metadata restore on the target cluster because pg_restore will
error trying to create a view using types that do not exist anymore.
This is not ideal as we could be many hours into upgrade before a
failure occurs. In order to detect views with removed types. A
query_tree_walker is used to walk the nodes of all views. Each node is
then checked to see if it is a type. If it is, we then check if the oid
of the type matches one that is not present in GPDB7.

Currently we only notify end users of views that contain a removed type.
We do not log what specific part of the view triggered the check for a
removed type.