Prevent overwriting of worker in ORCA (#17096)
In some cases, during optimization of queries with call to a function, it is observed that another instance of ORCA optimization is launched.

At the start of optimization, a new worker is registered in the 'Worker Pool Manager'. Currently ORCA supports only one worker, if a new worker is registered on top of an existing worker, then the details of the old worker is lost, causing undefined behaviour. Thus, through this PR, aim is to stop overwriting of workers. Similar work has been done in PR #17097.

Now, before registering a new worker it is first verified that no registered worker exists. If any worker exists, then new worker is not registered and ORCA falls back to planner for this optimization request. One known scenario in which ORCA tries to call itself is during static partition elimination (SPE), thus during SPE, ORCA is disabled so that call is made only to Planner.
---------

Co-authored-by: nishant sharma <nishants4@nishants46M6RY.vmware.com>