Handle Invalid Return Value For Opfamily Mdid. (#17229)
It was observed that while populating default opfamilies in PopulateDefaultOpfamilies(), for a data type the retrieved opfamily can be 'InvalidOid'. Eg - for 'json', the distribution opfamily is 'InvalidOid'. Using this invalid value for further processing leads to crash in ORCA. To prevent this condition, a check is added before using the mdid of opfamily.

Co-authored-by: nishant sharma <nishants4@nishants46M6RY.vmware.com>