pg_upgrade: add func to check if view func with changed signature
Views that use functions with changed signature will cause upgrade to
fail. This happens during metadata restore on the target cluster
because pg_restore will error trying to create a view using functions
that have changed signature, and would fail since these functions won't
have the same args/rets anymore.
This is not ideal as we could be many hours into upgrade before
a view using a function with changed signature causes pg_upgrade to fail
In order to detect views with functions with changed signatures,
a query_tree_walker is used to walk the nodes of all views.
Each node is then checked to see if it is a function.
If it is, we then check if the function signature is same as the one
present in GPDB7

Currently we only notify end users of views that contain a function with
changed signature.
We do not log what specific part of the view triggered the check for a
removed function.