[6X] Default to log_checkpoints=on
Backported from 7X commit: ff2d5ba9bed22e71d20503e6689fb380160a480e,
with minor conflicts. Original 7X commit message follows:

Backported from upstream commit:
64da07c41a8c0a680460cdafc79093736332b6cf
with one important difference: we didn't absorb the
log_autovacuum_min_duration change. This is because in GPDB, autovacuum
driven autoanalyze can take longer due to the size of GPDB tables (and
also analyze is dispatched from the QD). So, we decide to be
conservative about that change.

Original commit message follows:

The idea here is that when a performance problem is known to have
occurred at a certain point in time, it's a good thing if there is
some information available from the logs to help figure out what
might have happened around that time.

This change attracted an above-average amount of dissent, because
it means that a server with default settings will produce some amount
of log output even if nothing has gone wrong. However, by my count,
the mailing list discussion had about twice as many people in favor
of the change as opposed. The reasons for believing that the extra
log output is not an issue in practice are: (1) the rate at which
messages can be generated by this setting is bounded to one every
few minutes on a properly-configured system and (2) production
systems tend to have a lot more junk in the log from that due to
failed connection attempts, ERROR messages generated by application
activity, and the like.

Bharath Rupireddy, reviewed by Fujii Masao and by me. Many other
people commented on the thread, but as far as I can see that was
discussion of the merits of the change rather than review of the
patch.

Discussion: https://postgr.es/m/CALj2ACX-rW_OeDcp4gqrFUAkf1f50Fnh138dmkd0JkvCNQRKGA@mail.gmail.com

Reviewed-by: Marbin Tan <tmarbin@vmware.com>