Fix ZSTD for RLE when specified at table level
When we backported ZSTD support for RLE from 7X, we missed one more
validation site to fix for the compresslevel check:
validate_and_adjust_options (N.A. to 7X).

This was leading to issues when compresslevels 5 and 6 were being
specified at the table-level. For instance:

postgres=# create table foo(i int) with (appendonly=true, orientation=column,
    compresstype=rle_type, compresslevel=6);
ERROR:  compresslevel=6 is out of range for rle_type (should be in the range 1 to 4)

Reviewed-by: Huansong Fu <fuhuansong@gmail.com>
Reviewed-by: Ashwin Agrawal <aashwin@vmware.com>