Use full precision for statistics values in Orca
Previously, if a frequency or ndv value was very small, it could get
truncated to 0. In 7X, we bypass this translation logic for queries in
the database, but in 6X we'll still do this translation. By translating
the full precision, we'll improve cardinality estimates for queries with
very small ndv or frequency values.

Note: CFilterCardinalityTest was modified to not apply the filter twice
as it introduced flakiness.