Disable Orca refcount test on Mac (#17404)
This test ensures that an exception is raised with an illegal refcount.
However, MacOS seems to have extra protection and no exception is raised.

GPOS_DELETE_ARRAY indirectly calls into CMemoryPoolTracker::DeleteImpl()
which sets memory to 0xcdcdcdcd. But the last step in that function
calls clib::Free(header) which on MacOS unexpectedly set the bits back
to 0. The header is part of the allocation block so that's probably
legal.  Although Mac sets this to 0, it's not necessarily something we
can assume.

(cherry picked from commit 31586b01d0a8064e164d23cb04546691868d396f)