Fix mdqa bug as groupkey is constant (#17377)
we should distinguish between an empty groupClause and a constant one,
as they are clearly different. Within the function cdb_grouping_planner,
when root->group_pathkeys == NULL and root->parse->groupClause
aligns with the subpath's locus, it indicates a constant groupClause in
the query. Otherwise, the groupClause should be empty.