Fix pg_event_trigger_ddl_commands (#269)
The function issued the "invalid non-zero objectSubId for object class" error
when an event trigger fired after the following utility query types:
T_IndexStmt, T_CommentStmt, T_AlterExtensionContentsStmt, T_CreateTrigStmt,
T_AlterTypeStmt, T_CreateExternalStmt, T_DefineStmt, T_AlterDomainStmt (see
tests from the patch). The reason of the error was that the address variable had
not been initialized before it was added to the list in
EventTriggerCollectSimpleCommand.

Add initialization for commandCollected, because the variable was never set to
false, but EventTriggerCollectSimpleCommand should be called when
commandCollected is false.
Add initialization for address and secondaryObject to make sure that the same
error will never happen. When empty address is added to the list, this address
is skipped without error.