pgaudit: Add logging of AST for CTAS (#271)
Add the ast_ctas value for the pgaudit.log GUC.
When ast_ctas is included in pgaudit.log and the CREATE TABLE AS SELECT query is
executed, then the abstract syntax tree of the query's SELECT part is logged.
The tree is logged after QueryRewrite to replace views with tables. The tree is
useful to understand where new table columns data is taken from.