pgaudit: Add logging of AST for SELECTs (#272)
Add the ast_sel value for the pgaudit.log GUC.
When ast_sel is included in pgaudit.log and a SELECT query is executed, then
the abstract syntax tree of the query is logged.
The tree is logged before planning. The tree is useful to find out which columns
from which tables are used in SELECTs.