Add a copy of ic-good-opt-on, but with pgaudit (#274) (#275)
When the ic-good-opt-on tests were run with pgaudit, some of them failed because
pgaudit output warnings "unrecognized node type". To make the tests comportable
with pgaudit add cases for Greenplum specific node types to GetCommandLogLevel.
GetCommandLogLevel is called only from check_log_statement which returns whether
to log a query. Queries of added types are logged as DDL.