pgaudit: Add logging of AST for INSERT, UPDATE, DELETE
Add the ast_mod value for the pgaudit.log GUC.
When ast_mod is included in pgaudit.log then the query's abstract syntax tree is
logged for INSERT, UPDATE, DELETE.
The tree is logged before planning. The tree is useful to find out where table
columns data is taken from.
Add the stack_push_cmd function to initialize a stack item in planner_hook in
the same way as in ExecutorStart_hook.