Fix AOCO tables expansion (#318)
When a column was added to AOCO after adding new segments to cluster and before
the table expansion, then an error happened on gpexpand calling pg_relation_size
because pg_aocsseg.vpinfo contains invalid value. The table was not expanded in
this case.

A new segment is created based on master using pg_basebackup. AOCO tables on
master are empty, but their pg_aocsseg table contains rows where tupcount > 0,
but all EOFs in vpinfo are 0. These rows are copied to new segments. When
a column is added then the vpinfo column is not updated on the new segment,
because the aocs_addcol_emptyvpe function changes vpinfo when tupcount is 0
and AOCSFileSegInfoAddVpe is called from other functions when EOFs in vpinfo are
greater than zero.

GPDB 7 is not affected, because pg_aocsseg on master does not contain rows.

We cannot change data in pg_aocsseg on master, but we can truncate AOCO tables
on new segments.

Add cleanup to the gpctas test because the sro_mv_issue_11999 materialized view
is not refreshable. So we got an expected error running gpexpand in bfv_gpexpand.